Refactor: modify constants
