+++ client/src/constants/index.ts
@@ -1,2 +1,2 @@
 export const API_END_POINT = 'http://localhost:8080';
-export const USER_TOKEN = 'user token';
+export const USER_TOKEN_KEY = 'user token';
