+++ client/src/index.tsx
@@ -2,9 +2,8 @@ import axios from 'axios';
 import React from 'react';
 import { createRoot } from 'react-dom/client';
 
-import { API_END_POINT } from './constants/constants';
-
 import { App } from '@/App';
+import { API_END_POINT } from '@/constants';
 
 const root = createRoot(document.getElementById('root') as HTMLDivElement);
 axios.defaults.baseURL = API_END_POINT;
