+++ client/src/layout/index.tsx
@@ -3,14 +3,14 @@ import { Outlet } from 'react-router-dom';
 
 import { NavBar } from './NavBar';
 
-import { USER_TOKEN } from '@/constants/constants';
+import { USER_TOKEN_KEY } from '@/constants';
 import useMovePage from '@/hooks/useMovePage';
 
 export const Layout = () => {
   const [goLogin] = useMovePage('/login');
 
   useEffect(() => {
-    const userToken = localStorage.getItem(USER_TOKEN);
+    const userToken = localStorage.getItem(USER_TOKEN_KEY);
     if (!userToken) {
       alert('로그인이 필요합니다. 로그인 페이지로 이동합니다.');
       goLogin();
